Extract generate data
